requirements.txt
Pillow==5.3.0
pytest==4.0.0
